refactor(vue): restructured components and views responsible dirs
