+++ resources/components/administration/NavFooter.vue
