+++ resources/views/settings/Appearance.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { Head } from '@inertiajs/vue3';
 
-import AppearanceTabs from '@/components/AppearanceTabs.vue';
 import HeadingSmall from '@/components/HeadingSmall.vue';
 import { type BreadcrumbItem } from '@/types';
+import AppearanceTabs from './partial/AppearanceTabs.vue';
 
 import SettingsLayout from '@/layouts/settings/Layout.vue';
